change to new timeout method and cleanup

+++ tornado-chat.py
@@ -4,7 +4,6 @@ import tornado.websocket
 from tornado.escape import json_decode
 import os, string, random, json
 
-MAX_ROOMS = 50
 BUFFER_SIZE = 100
 DEFAULT_PORT = 3000
 MAX_CT = 512
@@ -16,83 +15,49 @@ LOG_DIRECTORY = "/tmp/cryptoneko"
 class MessagingBuffer:
     
     def __init__(self, code=None):
-        self.limit = BUFFER_SIZE
-        self.message_buffer = [None for i in range(self.limit)]
-        self.index = 0
-        self.current = 0
+        self.initialized = False
+        self.log_file = f"{LOG_DIRECTORY}/{code}.log"
 
-        # here is where we will try to just add in a file system 
-        self.log_file = f"{code}.log"
-
-        with open(self.log_file, "w") as writeit:
-            pass 
-
-
-    def _ordered_yield(self):
-        for i in range(self.limit):
-            x = (self.index + i) % self.limit
-            if self.message_buffer[x]:
-                yield self.message_buffer[x]
+    def init_log(self):
+        if not os.path.isfile(self.log_file):
+            with open(self.log_file, "w") as writeit:
+                pass 
+        self.initialized = True
 
     def push(self, msg):
-        # self.message_buffer[self.index] = msg 
-        # self.index = (self.index + 1) % self.limit
-        # self.current += 1
+        if not self.initialized:
+            self.init_log()
         with open(self.log_file, "a") as log:
-            print(msg)
             log.write(json.dumps(msg) + "\n")
 
     def get_all(self):
-        #return list(self._ordered_yield())
+        if not self.initialized:
+            self.init_log()
         with open(self.log_file, "r") as log:
             raw = log.readlines()
-            print("debug this!")
-            print(raw)
             return [json.loads(x) for x in raw]
 
 class RoomQueue:
 
     def __init__(self):
         self.current = 0
-        self.map = {}
-        self.rooms = []
-        # for i in range(MAX_ROOMS):
-        #     self.rooms.append( None )
+        self.rooms = {}
 
-        # refactor all this crap to not use the arrays and other structures 
-        # I think all we really need here is a dictionary of the current rooms 
-        # so we put the code as the key and then add the buffer as the content 
-        # and we don't need anything else 
-        # lets see if we can use linux log rotation utilities to handle the chat 
-        # logs and manage the number of files
         if not os.path.exists(LOG_DIRECTORY):
-            os.makedirs(directory)
+            os.makedirs(LOG_DIRECTORY)
         log_files = os.listdir(LOG_DIRECTORY)
         for f in log_files:
-            print(f)
-            if f.endswith(".log")
-            self.rooms.append()
+            if f.endswith(".log"):
+                code = f[:-4]
+                self.rooms[code] = MessagingBuffer(code=code)
 
     def room_code(self):
         return ''.join(random.choice(string.ascii_letters) for x in range(5))
 
     def new_room(self):
-        
         code = self.room_code()
-        self.rooms[self.current] = MessagingBuffer(code=code)
-        for key, index in self.map.items():
-            if self.current == index:
-                self.map.pop(key)
-                break
-        
-        
-        self.map[code] = self.current
-        self.current = (self.current + 1) % MAX_ROOMS
-     
+        self.rooms[code] = MessagingBuffer(code=code)
         return code
-
-    def room_index(self, code):
-        return self.map.get(code, None)
 
 
 class MainHandler(tornado.web.RequestHandler):
@@ -105,8 +70,8 @@ class RoomHandler(tornado.web.RequestHandler):
 
     def get(self, code):
         nick = self.get_argument("nick", DEFAULT_NICK)
-        index = self.application.room_queue.room_index(code)
-        if index == None:
+        queue = self.application.room_queue.rooms.get(code)
+        if not queue:
             raise tornado.web.HTTPError(
                 status_code=404,
                 reason="Room Not Found, nya"
@@ -129,8 +94,8 @@ class EchoWebSocket(tornado.websocket.WebSocketHandler):
 
     def on_message(self, message):
         msg = json_decode(message)
-        buffer = self.application.room_queue.rooms[self.application.room_queue.room_index(self._code)]
-            
+        buffer = self.application.room_queue.rooms.get(self._code)
+
         if 'ciphertext' in msg and len(msg.get('ciphertext')) < MAX_CT:
             buffer.push(msg)
             self.broadcast(self._code, {"messages": [msg]} )
@@ -144,7 +109,7 @@ class EchoWebSocket(tornado.websocket.WebSocketHandler):
             self.application.room_set[code] = set()
         self.application.room_set.get(code).add(self)
 
-        buffer =  self.application.room_queue.rooms[self.application.room_queue.room_index(code)]
+        buffer =  self.application.room_queue.rooms[code]
         self.write_message({'messages':buffer.get_all()})
